backend/topology.py
from google.cloud import asset_v1
import os

class TopologyEngine:
    def __init__(self):
        self.client = asset_v1.AssetServiceClient()
        self.project_id = os.getenv("PROJECT_ID", "local-project")

    def get_service_dependencies(self, resource_name: str) -> list[str]:
        """
        Mock implementation for Phase 3. 
        In production, this would use SearchAllResources or SearchAllIamPolicies.
        """
        # Logic to query Cloud Asset Inventory
        # For prototype, we infer based on naming or mock returns for specific resources
        
        dependencies = []
        
        # Example mock logic
        if "cloud-run" in resource_name:
            dependencies.append(f"//pubsub.googleapis.com/projects/{self.project_id}/topics/sre-incident-bus")
            dependencies.append(f"//logging.googleapis.com/projects/{self.project_id}/logs/run.googleapis.com%2Fstderr")
        
        # Actual API call (commented out for now as we might not have real infra to query yet)
        # scope = f"projects/{self.project_id}"
        # request = asset_v1.SearchAllResourcesRequest(scope=scope, query=f"name:{resource_name}")
        # page_result = self.client.search_all_resources(request=request)
        # for resource in page_result:
        #     ...
            
        return dependencies
